LPS-76509 Move MBTestUtil to module and deprecate the one in core
