+++ modules/apps/collaboration/message-boards/message-boards-test-util/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Message Boards Test Utilities
+Bundle-SymbolicName: com.liferay.message.boards.test.util
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.message.boards.test.util
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Message Boards
