+++ modules/apps/collaboration/message-boards/message-boards-test-util/build.gradle
@@ -1,0 +1,8 @@
+dependencies {
+	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.53.0"
+	provided group: "com.liferay.portal", name: "com.liferay.portal.test", version: "2.0.0"
+}
+
+liferay {
+	deployDir = file("${liferayHome}/osgi/test")
+}
