+++ modules/apps/collaboration/message-boards/message-boards-test-util/src/main/resources/com/liferay/message/boards/test/util/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
