+++ modules/apps/collaboration/message-boards/message-boards-test/build.gradle
@@ -5,6 +5,7 @@ dependencies {
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.0.0"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	testIntegrationCompile project(":apps:collaboration:message-boards:message-boards-api")
+	testIntegrationCompile project(":apps:collaboration:message-boards:message-boards-test-util")
 	testIntegrationCompile project(":apps:collaboration:social:social-activity-test-util")
 	testIntegrationCompile project(":apps:foundation:portal-search:portal-search-test-util")
 	testIntegrationCompile project(":apps:web-experience:export-import:export-import-test-util")
